+++ x/evm/keeper/benchmark_test.go
@@ -12,6 +12,7 @@ import (
 
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	ethermint "github.com/tharsis/ethermint/types"
+	"github.com/tharsis/ethermint/x/evm/keeper"
 	"github.com/tharsis/ethermint/x/evm/types"
 )
 
@@ -94,4 +95,41 @@ func BenchmarkTokenMint(b *testing.B) {
 		nonce := suite.app.EvmKeeper.GetNonce(suite.address)
 		return types.NewTx(suite.app.EvmKeeper.ChainID(), nonce, &contract, big.NewInt(0), 410000, big.NewInt(1), nil, nil, input, nil)
 	})
+}
+
+func DoBenchmarkDeepContextStack(b *testing.B, depth int) {
+	begin := []byte{0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00}
+	end := []byte{0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff}
+
+	suite := KeeperTestSuite{}
+	suite.DoSetupTest(b)
+
+	transientKey := suite.app.GetTKey(types.TransientKey)
+
+	var stack keeper.ContextStack
+	stack.Reset(suite.ctx)
+
+	for i := 0; i < depth; i++ {
+		stack.Snapshot()
+
+		store := stack.CurrentContext().TransientStore(transientKey)
+		store.Set(begin, []byte("value"))
+	}
+
+	store := stack.CurrentContext().TransientStore(transientKey)
+	for i := 0; i < b.N; i++ {
+		store.Iterator(begin, end)
+	}
+}
+
+func BenchmarkDeepContextStack1(b *testing.B) {
+	DoBenchmarkDeepContextStack(b, 1)
+}
+
+func BenchmarkDeepContextStack10(b *testing.B) {
+	DoBenchmarkDeepContextStack(b, 10)
+}
+
+func BenchmarkDeepContextStack13(b *testing.B) {
+	DoBenchmarkDeepContextStack(b, 13)
 }